Move SponsorsTable from Dashboard to Sponsors page
